requirements.txt
django~=3.2.5
django-ckeditor
django-cors-headers
django-easy-select2
django-extensions
django-jazzmin
django-js-asset
django-ninja
django-silk
pillow
typing-extensions
pydantic[email]~=1.8.2
pretty-errors
python-jose

cryptography~=3.4.7
twilio~=6.62.1